+++ package.json
@@ -3,12 +3,13 @@
   "version": "1.4.0",
   "description": "A package to help you find out which of your `node_modules` aren't ES5 so you can add them to your transpilation steps.",
   "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "repository": "https://github.com/obahareth/are-you-es5.git",
   "author": "Omar Bahareth <omar@omar.engineer>",
   "license": "MIT",
   "private": false,
   "bin": {
-    "are-you-es5": "./dist/index.js"
+    "are-you-es5": "./dist/cli.js"
   },
   "dependencies": {
     "acorn": "^6.0.6",
