+++ tsconfig.json
@@ -7,11 +7,8 @@
     "noImplicitAny": true,
     "moduleResolution": "node",
     "sourceMap": true,
+    "declaration": true,
     "outDir": "dist"
   },
-
-
-  "include": [
-    "src/**/*"
-  ]
+  "include": ["src/**/*"]
 }
